AbstractFollower is its own file now
