+++ src/server/game/Entities/Unit/Unit.cpp
@@ -17,6 +17,7 @@
  */
 
 #include "Unit.h"
+#include "AbstractFollower.h"
 #include "Battlefield.h"
 #include "BattlefieldMgr.h"
 #include "Battleground.h"
